+++ Sources/KeyboardKit/Appearance/KeyboardAppearance.swift
@@ -63,7 +63,7 @@ public protocol KeyboardAppearance: AnyObject {
      The button style to use for a certain `action`, given a
      certain `isPressed` state.
      */
-    func buttonStyle(for action: KeyboardAction, isPressed: Bool) -> KeyboardButtonStyle
+    func buttonStyle(for action: KeyboardAction, isPressed: Bool) -> KeyboardStyle.Button
     
     /**
      The button text to use for a certain `action`, if any.