+++ Sources/KeyboardKit/_Deprecated/KeyboardStyle+Deprecated.swift
@@ -3,6 +3,15 @@ import Foundation
 @available(*, deprecated, renamed: "KeyboardStyle.ActionCallout")
 public typealias KeyboardActionCalloutStyle = KeyboardStyle.ActionCallout
 
+@available(*, deprecated, renamed: "KeyboardStyle.Button")
+public typealias KeyboardButtonStyle = KeyboardStyle.Button
+
+@available(*, deprecated, renamed: "KeyboardStyle.ButtonBorder")
+public typealias KeyboardButtonBorderStyle = KeyboardStyle.ButtonBorder
+
+@available(*, deprecated, renamed: "KeyboardStyle.ButtonShadow")
+public typealias KeyboardButtonShadowStyle = KeyboardStyle.ButtonShadow
+
 @available(*, deprecated, renamed: "KeyboardStyle.Callout")
 public typealias KeyboardCalloutStyle = KeyboardStyle.Callout